Split push dispatcher into focused submodules.
Extract delivery health and in-flight task tracking from the monolithic
dispatcher so the queue/worker logic stays easier to navigate and bisect.
